### 2021-03-14 #3106 @whedon
[PRE REVIEW]: BIDSonym - a BIDSapp for the pseudo-anonymization of neuroimaging datasets
**Submitting author:** @peerherholz (<a href="http://orcid.org/0000-0002-9840-6257">Peer Herholz</a>)
**Repository:** <a href="https://github.com/PeerHerholz/BIDSonym/" target ="_blank">https://github.com/PeerHerholz/BIDSonym/</a>
**Version:** v0.0.4
