### 2021-03-14 #3106 @PeerHerholz
Hi @arfon,

thx for the pointer and update.
Sorry for the obvious error...I compiled it locally via the `docker image` and also the `web version`, but overlooked my missing `orcid` both times. I updated the `paper.md` and added my `orcid`.
Re editors: of course, I understand, thx for letting me know.

Cheers, Peer
