### 2021-03-15 #3106 @arfon
> pandoc: Bidsonym.bib: openBinaryFile: does not exist (No such file or directory)

@PeerHerholz - Pandoc is case sensitive for filenames your BibTeX file has a different case from how you're referencing it in the `paper.md`.
