### 2021-04-05 #3106 @danielskatz
@PeerHerholz - this is waiting for your fix as described by @arfon above - https://github.com/openjournals/joss-reviews/issues/3106#issuecomment-799210422
