+++ Dockerfile
@@ -111,134 +111,8 @@ RUN bash -c 'chmod a+x /home/bm/bidsonym/fs_data/mri_deface'
 
 RUN bash -c 'source activate bidsonym && cd /home/bm && pip install -e .'
 
-COPY ["example_data", "/home/bidsonym/example_data"]
-
 ENV IS_DOCKER="1"
 
 WORKDIR /tmp/
 
 ENTRYPOINT ["/neurodocker/startup.sh", "bidsonym"]
-
-RUN echo '{ \
-    \n  "pkg_manager": "apt", \
-    \n  "instructions": [ \
-    \n    [ \
-    \n      "base", \
-    \n      "neurodebian:stretch-non-free" \
-    \n    ], \
-    \n    [ \
-    \n      "install", \
-    \n      [ \
-    \n        "fsl-core", \
-    \n        "git", \
-    \n        "num-utils", \
-    \n        "gcc", \
-    \n        "g++", \
-    \n        "curl", \
-    \n        "build-essential", \
-    \n        "nano" \
-    \n      ] \
-    \n    ], \
-    \n    [ \
-    \n      "run_bash", \
-    \n      "curl -sL https://deb.nodesource.com/setup_10.x | bash - && apt-get install -y nodejs && apt-get install -y npm" \
-    \n    ], \
-    \n    [ \
-    \n      "add_to_entrypoint", \
-    \n      "source /etc/fsl/fsl.sh" \
-    \n    ], \
-    \n    [ \
-    \n      "miniconda", \
-    \n      { \
-    \n        "conda_install": [ \
-    \n          "python=3.6", \
-    \n          "numpy", \
-    \n          "nipype", \
-    \n          "nibabel", \
-    \n          "pandas" \
-    \n        ], \
-    \n        "pip_install": [ \
-    \n          "deepdefacer", \
-    \n          "tensorflow", \
-    \n          "scikit-image" \
-    \n        ], \
-    \n        "create_env": "bidsonym", \
-    \n        "activate": true \
-    \n      } \
-    \n    ], \
-    \n    [ \
-    \n      "run_bash", \
-    \n      "source activate bidsonym && git clone https://github.com/poldracklab/pydeface.git && cd pydeface && python setup.py install && cd -" \
-    \n    ], \
-    \n    [ \
-    \n      "run_bash", \
-    \n      "source activate bidsonym && git clone https://github.com/nipy/quickshear.git  && cd quickshear && python setup.py install && cd -" \
-    \n    ], \
-    \n    [ \
-    \n      "run_bash", \
-    \n      "source activate bidsonym && git clone https://github.com/neuronets/nobrainer.git  && cd nobrainer && python setup.py install && cd -" \
-    \n    ], \
-    \n    [ \
-    \n      "run_bash", \
-    \n      "mkdir -p /opt/nobrainer/models && cd /opt/nobrainer/models && curl -LJO  https://github.com/neuronets/nobrainer-models/releases/download/0.1/brain-extraction-unet-128iso-model.h5 && cd ~ " \
-    \n    ], \
-    \n    [ \
-    \n      "run_bash", \
-    \n      "git clone https://github.com/mih/mridefacer" \
-    \n    ], \
-    \n    [ \
-    \n      "env", \
-    \n      { \
-    \n        "MRIDEFACER_DATA_DIR": "/mridefacer/data" \
-    \n      } \
-    \n    ], \
-    \n    [ \
-    \n      "run_bash", \
-    \n      "npm install -g bids-validator@1.5.4" \
-    \n    ], \
-    \n    [ \
-    \n      "run_bash", \
[…]
