+++ bidsonym/run_deeid.py
@@ -100,7 +100,7 @@ def run_deeid():
         if part not in layout.get_subjects():
             list_part_prob.append(part)
     if len(list_part_prob) >= 1:
-        raise Exception("The participant(s) you indicated are present in the BIDS dataset, please check again."
+        raise Exception("The participant(s) you indicated are not present in the BIDS dataset, please check again."
                         "This refers to:")
         print(list_part_prob)
 
@@ -135,7 +135,7 @@ def run_deeid():
                 run_brain_extraction_nb(T1_file, subject_label, args.bids_dir)
 
             check_meta_data(args.bids_dir, subject_label, list_check_meta)
-            source_t1w = copy_no_deid(subject_label, args.bids_dir, T1_file)
+            source_t1w = copy_no_deid(args.bids_dir, subject_label, T1_file)
 
             if args.del_meta:
                 del_meta_data(args.bids_dir, subject_label, list_field_del)
