### 2021-04-12 #3169 @oliviaguest
Hi all! 👋 Thank you so much, @deep-introspection, @chrisgorgo, @neuromusic for accepting to review this. Please read the instructions above. If you have trouble ticking things off in the list above, remember you need to click to "accept" here: https://github.com/openjournals/joss-reviews/invitations! Any questions, feedback on the paper, etc., please post here. Any very code-specific questions, suggestions, etc., please use the issues in the code repo and link to them from this thread, so we can all keep track of them.

For examples of how this process plays out feel free to skim previous reviews, such as: #2285 and #2348. ☺️
