### 2021-04-13 80ae882
singularity pull from docker
Trying to do it from shub yields " While pulling shub image: error fetching image to cache: failed to get manifest for: shub://PeerHerholz/BIDSonym: the requested manifest was not found in singularity hub".  Doing it from Docker works.
