### 2021-04-13 eb1f0c3
test updates wrt container behavior
This PR removes the generation of a Singularity recipe given problems wrt the `--copy` argument in `neurodocker`. Furthermore, it adds a new branch that aims to host tests for future updates concerning the container behavior, as well as adds  the printing the list of meta-data files that will be evaluated.
