### 2021-04-13 08a233d
adapt graphics function wrt --deface_t2w argument
