### 2021-04-13 4ca3ed9
update plotting call in main function
This commit should resolve issues related to the generation of the QC plots, specifically how the function is called under which circumstances.
