+++ bidsonym/run_deeid.py
@@ -168,17 +168,17 @@ def run_deeid():
                 elif args.brainextraction == 'nobrainer':
                     run_brain_extraction_nb(T2_file, subject_label, args.bids_dir)
 
-                source_t2w = copy_no_deid(subject_label, args.bids_dir, T2_file)
+                source_t2w = copy_no_deid(args.bids_dir, subject_label, T2_file)
                 run_t2w_deface(source_t2w, T1_file, T2_file)
 
         rename_non_deid(args.bids_dir, subject_label)
 
-        if sessions_to_analyze is None and args.deface_t2w is False:
+        if not sessions_to_analyze and args.deface_t2w is False:
             create_graphics(args.bids_dir, subject_label, session=None, t2w=None)
         elif sessions_to_analyze and args.deface_t2w is False:
             for session in sessions_to_analyze:
                 create_graphics(args.bids_dir, subject_label, session=session, t2w=None)
-        elif sessions_to_analyze is None and args.deface_t2w:
+        elif not sessions_to_analyze and args.deface_t2w:
             create_graphics(args.bids_dir, subject_label, session=None, t2w=True)
         elif sessions_to_analyze and args.deface_t2w:
             for session in sessions_to_analyze:
