### 2021-04-14 573011c
remove singularity from generate_images.sh and update function
