### 2021-05-20 PeerHerholz/BIDSonym#53 @neuromusic
minor typos and errors in JOSS manuscript
related to JOSS Review at https://github.com/openjournals/joss-reviews/issues/3169

- typo, line 16 of draft PDF: "par-ticipant" ➡️ "participant"
- typo, line 17 of draft PDF: "Re-view" ➡️ "Review"
- typo, line 45 of draft PDF: "pypi" ➡️ "PyPI"
- error, line 53 of draft PDF: "??"
- typo, line 91 of draft PDF: "Of" ➡️ "of"
