### 2021-05-21 PeerHerholz/BIDSonym#54 @neuromusic
improvements to installation documentation
(related to JOSS Review at openjournals/joss-reviews#3169)

the installation instructions suggest `docker pull peerherholz/bidsonym:version` to install, however "version" is invalid. this should be updated to a command that can be copied and pasted, like `docker pull peerherholz/bidsonym` or `docker pull peerherholz/bidsonym:latest`

similarly, there's a typo in the markup for the code to pull from singularity at https://peerherholz.github.io/BIDSonym/installation.html#directly-pulling-from-singularity-hub
