### 2021-05-21 PeerHerholz/BIDSonym#55 @neuromusic
tests?
(related to JOSS Review at openjournals/joss-reviews#3169)

The JOSS review checklist asks "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"

I can see that there's some travis configuration that does some linting and verifies the Python package builds and Docker container builds, but I'm struggling to find any unit tests. Do you any have additional automated tests?
