### 2021-05-27 9cfbf43
put usage msg in code-block for better legibility
