### 2021-06-07 a629ae2
JOSS review - typos & errors in manuscript
This commit and PR addresses and fixes #53, more precisely typos and errors as outlined by @neuromusic. Thx a lot for the feedback!
