### 2021-06-07 PeerHerholz/BIDSonym#53 @PeerHerholz
Hi @neuromusic,

thanks so much for the feedback. I addressed and fixed these issues via [a629ae2](https://github.com/PeerHerholz/BIDSonym/commit/a629ae2cb2ac0b9d914a7ecd00cccd936d35aef6). I'll recreate the updated PDF in the [review thread](https://github.com/openjournals/joss-reviews/issues/3169).
