### 2021-06-07 #3169 @PeerHerholz
Ahoi hoi everyone,

I just addressed and fixed typos and errors in the manuscript outlined by @neuromusic in this [issue](https://github.com/PeerHerholz/BIDSonym/issues/53).  I'll regenerate the PDF to reflect the respective changes.
