### 2021-06-07 PeerHerholz/BIDSonym#54 @PeerHerholz
Ahoi hoi @neuromusic,

thanks a lot for the feedback. You're definitely right that only including `docker pull peerherholz/bidsonym:version` will lead to problems. However, to also prevent `copy paste errors/problems` via including only certain examples for specific versions (e.g.  `docker pull peerherholz/bidsonym:version`), I extended the respective part(s) of the documentation to include and explain both the `place holder` and the `specific version` option. Furthermore, the section related to `Singularity` was updated and `markup` related problems fixed.

Thanks again!
