### 2021-06-07 ee00478
JOSS review - improvements to installation documentation
This commit and PR addresses and fixes #54, entailing improvements to the installation section of the documentation as outlined by @neuromusic:

- include and explain both place holder and specific version example wrt `docker pull`
- update `Singularity` section
- fix markup problems
