+++ docs/source/citing.rst
@@ -5,5 +5,9 @@ Citing *BIDSonym*
 =================
 
 If you are using ``BIDSonym`` in your work, please consider citing the 
-`GitHub repository <https://github.com/PeerHerholz/BIDSonym>`_ 
-until we have a formal publication out there.
+`preprint <https://psyarxiv.com/3aknq/>`_:
+
+Herholz, P., Ludwig, R. M., & Poline, J. (2021, January 24). BIDSonym - a BIDSapp for the pseudo-anonymization of neuroimaging datasets. 
+https://doi.org/10.31234/osf.io/3aknq
+
+and specify the exact ``version`` you used, ideally including the respective ``DOI`` as generated through ``Zenodo``. 
