### 2021-06-07 006c691
Update docs after building Travis build 152 of PeerHerholz/BIDSonym
The docs were built from the branch 'master' against the commit
ee004788cffa14f76167b88314145b5738059022.

The Travis build that generated this commit is at
https://travis-ci.org/PeerHerholz/BIDSonym/jobs/773805953.

The doctr command that was run is

    /home/travis/virtualenv/python3.6.7/bin/doctr deploy --built-docs docs/build/html .
