+++ _sources/installation.rst.txt
@@ -7,14 +7,15 @@ Installation
 In general, there are two distinct ways to install and use ``BIDSonym``:
 either through virtualization/container technology, that is `Docker`_ or
 `Singularity`_, or in a `Bare metal version (Python 3.6+)`_.
-Using local container method is highly recommended.
-Once you are ready to run ```BIDSonym```, see `Usage <./usage.rst>`_ for details.
+Using a container method is highly recommended as they entail entire operating systems through kernel level virtualization and
+thus include all software necessary to run ``BIDSonym``, while at the same time presenting a lightweight alternative to virtual machines.
+Once you are ready to run ``BIDSonym``, see `Usage <./usage.rst>`_ for details.
 
 Docker
 ======
 
 In order to run ```BIDSonym``` in a Docker container, Docker must be `installed
-<https://docs.docker.com/engine/installation/>`_.
+<https://docs.docker.com/engine/installation/>`_ on your system.
 Once Docker is installed, you can get ``BIDSonym`` through running the following
 command in the terminal of your choice:
 
@@ -22,11 +23,26 @@ command in the terminal of your choice:
 
     docker pull peerherholz/bidsonym:version
 
+Where ``version`` is the specific version of ``BIDSonym`` you would like to use. For example, if you want 
+to employ the ``latest``/most up to date ``version`` you can either run 
+
+.. code-block:: bash
+
+    docker pull peerherholz/bidsonym:latest
+
+or the same command withouth the ``:latest`` tag, as ``Docker`` searches for the ``latest`` tag by default.
+However, as the ``latest`` version is subject to changes and not necessarily in synch with the most recent ``numbered version``, it 
+is recommend to utilize the latter to ensure reproducibility. For example, if you want to employ ``BIDSonym v0.0.4`` the command would look as follows:
+
+.. code-block:: bash
+
+    docker pull peerherholz/bidsonym:v0.0.4
+
 .. note::
 
    As of November 2020, `images older than 6 months will be deleted from Dockerhub
    <https://www.docker.com/pricing/retentionfaq>`_. As this is very problematic for everything
-   reproducibility and version control, every version of the BIDSonym images are additionally
+   reproducibility and version control, every version of the ``BIDSonym images`` are additionally
    uploaded on `OSF <https://osf.io/x4dku/>`_ and can be installed as outlined further below.
 
 After the command finished (it may take a while depending on your internet connection),
@@ -42,52 +58,43 @@ you can run ``BIDSonym`` like this:
         --participant_label label \
         --deid defacing_algorithm \
 
-Please have a look at the examples under Usage to get more information
+Please have a look at the examples under `Usage <./usage.rst>`_ to get more information
 about and familiarize yourself with ``BIDSonym``'s functionality.
 
 
 Singularity
 ===========
 
-For security reasons, many HPCs (e.g., TACC) do not allow Docker containers, but do
-allow `Singularity <https://github.com/singularityware/singularity>`_ containers.
-
-Directly pulling from Singularity Hub
-----------------------------------------------------------
-The ``BIDSonym`` Singularity image can directly be pulled from
-Singularity Hub via:
-
-    $ singularity pull PeerHerholz/BIDSonym
+For security reasons, many HPCs (e.g., TACC) do not allow Docker containers, but support
+allow `Singularity <https://github.com/singularityware/singularity>`_ containers. Depending
+on the ``Singularity`` version available to you, there are two options to get ``BIDSonym`` as
+a ``Singularity image``.
 
 Preparing a Singularity image (Singularity version >= 2.5)
[…]
