### 2021-06-07 5ddcc42
mridefacer fix and update
This commit and PR fixes and updates mridefacer, as it crashed based on inconsistencies in the respective function's setup. 
The following files were changed:

- generate_iamges.sh & Dockerfile -> FSL templates were added to the container as they are needed by mridefacer
- defacing_algorithms -> mridefacer commands and workflow were updated to remove problems wrt copying of the face mask & add the --outdir argument to write defaced images to the respective BIDS root directory
- run_deeid -> updated to reflect the changes made in defacing_algorithms
