+++ Dockerfile
@@ -1,6 +1,6 @@
 # Your version: 0.6.0 Latest version: 0.7.0
 # Generated by Neurodocker version 0.6.0
-# Timestamp: 2021-04-13 14:10:53 UTC
+# Timestamp: 2021-06-07 18:59:26 UTC
 # 
 # Thank you for using Neurodocker. If you discover any issues
 # or ways to improve this software, please submit an issue or
@@ -44,6 +44,7 @@ ENTRYPOINT ["/neurodocker/startup.sh"]
 RUN apt-get update -qq \
     && apt-get install -y -q --no-install-recommends \
            fsl-core \
+           fsl-mni152-templates \
            git \
            num-utils \
            gcc \
@@ -117,3 +118,121 @@ ENV IS_DOCKER="1"
 WORKDIR /tmp/
 
 ENTRYPOINT ["/neurodocker/startup.sh", "bidsonym"]
+
+RUN echo '{ \
+    \n  "pkg_manager": "apt", \
+    \n  "instructions": [ \
+    \n    [ \
+    \n      "base", \
+    \n      "neurodebian:stretch-non-free" \
+    \n    ], \
+    \n    [ \
+    \n      "install", \
+    \n      [ \
+    \n        "fsl-core", \
+    \n        "fsl-mni152-templates", \
+    \n        "git", \
+    \n        "num-utils", \
+    \n        "gcc", \
+    \n        "g++", \
+    \n        "curl", \
+    \n        "build-essential", \
+    \n        "nano" \
+    \n      ] \
+    \n    ], \
+    \n    [ \
+    \n      "run_bash", \
+    \n      "curl -sL https://deb.nodesource.com/setup_10.x | bash - && apt-get install -y nodejs && apt-get install -y npm" \
+    \n    ], \
+    \n    [ \
+    \n      "add_to_entrypoint", \
+    \n      "source /etc/fsl/fsl.sh" \
+    \n    ], \
+    \n    [ \
+    \n      "miniconda", \
+    \n      { \
+    \n        "conda_install": [ \
+    \n          "python=3.6", \
+    \n          "numpy", \
+    \n          "nipype", \
+    \n          "nibabel", \
+    \n          "pandas" \
+    \n        ], \
+    \n        "pip_install": [ \
+    \n          "deepdefacer", \
+    \n          "tensorflow", \
+    \n          "scikit-image" \
+    \n        ], \
+    \n        "create_env": "bidsonym", \
+    \n        "activate": true \
+    \n      } \
+    \n    ], \
+    \n    [ \
+    \n      "run_bash", \
+    \n      "source activate bidsonym && git clone https://github.com/poldracklab/pydeface.git && cd pydeface && python setup.py install && cd -" \
+    \n    ], \
+    \n    [ \
+    \n      "run_bash", \
+    \n      "source activate bidsonym && git clone https://github.com/nipy/quickshear.git  && cd quickshear && python setup.py install && cd -" \
+    \n    ], \
+    \n    [ \
+    \n      "run_bash", \
+    \n      "source activate bidsonym && git clone https://github.com/neuronets/nobrainer.git  && cd nobrainer && python setup.py install && cd -" \
+    \n    ], \
+    \n    [ \
+    \n      "run_bash", \
+    \n      "mkdir -p /opt/nobrainer/models && cd /opt/nobrainer/models && curl -LJO  https://github.com/neuronets/nobrainer-models/releases/download/0.1/brain-extraction-unet-128iso-model.h5 && cd ~ " \
+    \n    ], \
+    \n    [ \
+    \n      "run_bash", \
[…]
