### 2021-08-27 91e5e83
remove singularity related information
