+++ README.rst
@@ -28,10 +28,6 @@ BIDSonym
 .. image:: https://img.shields.io/docker/pulls/peerherholz/bidsonym
     :alt: Dockerpulls
     :target: https://cloud.docker.com/u/peerherholz/repository/docker/peerherholz/bidsonym
-    
-.. image:: https://www.singularity-hub.org/static/img/hosted-singularity--hub-%23e32929.svg
-    :alt: SingularityHub
-    :target: https://singularity-hub.org/collections/4645
 
 .. image:: https://img.shields.io/badge/License-BSD%203--Clause-blue.svg
     :alt: License
@@ -166,9 +162,6 @@ To get the BIDSonym Docker image, you need to `install docker <https://docs.dock
 
 :code:`docker pull peerherholz/bidsonym`
 
-To get its Singularity version, you need to `install singularity <https://singularity.lbl.gov/all-releases>`_ and within the terminal of your choice type:
-
-:code:`singularity pull docker://peerherholz/bidsonym`
 
 Documentation
 =============
