### 2021-08-27 1d2f48f
Update docs after building Travis build 156 of PeerHerholz/BIDSonym
The docs were built from the branch 'master' against the commit
91e5e838aea3a6e55f2ed66d6ce16e7697019945.

The Travis build that generated this commit is at
https://app.travis-ci.com/PeerHerholz/BIDSonym/jobs/534199652.

The doctr command that was run is

    /home/travis/virtualenv/python3.6.7/bin/doctr deploy --built-docs docs/build/html .
