+++ _sources/example.rst.txt
@@ -1,0 +1,132 @@
+------------------
+An usage example
+------------------
+
+In order to provide interested users the opportunity to test ``BIDSOnym`` and thus evaluate if they want to utilize 
+it within their own datasets, we provide an example step-by-step walkthrough based on an open non-deidentified datasets.
+
+**1. The example dataset**
+
+Finding non-deidentified open datasets is hard and rightfully so, because the privacy of the participants needs to be protected.
+Luckily for ``BIDSonym``, there a few ones out there where participants (mostly they were neuroscientits themselves) provided their own data in a 
+non-deidentified manner. For this example, we decided to go with an adapted version of the `MyConnectome dataset <http://myconnectome.org/wp/>`_
+which can be found on `OpenNeuro <https://openneuro.org/datasets/ds000031/versions/00001>`_, licensed under `PDDL <https://opendatacommons.org/licenses/pddl/summary/>`_.
+
+As it is quite a big dataset and for ease of use, we adapted the original dataset to only include data from one session (``ses-006``) and
+added some fake identifiers to the ``.json sidecar files`` of the ``T1w`` image to illustrate the pseudonymization of the ``metadata information``.
+You can download this adapted version of the dataset (~1.3 GB) `here <https://www.dropbox.com/s/ztsytlns8trlw3p/ds000031.zip?dl=1>`_.  
+
+After downloading it, please unzip the folder and place it under your preferred path. For this example, we will
+assume it is placed under ``/Users/peerherholz/Desktop`` which is the path to the desktop of my (Peer) local machine.
+Please remember to change the path to wherever you stored the example dataset!
+
+Using the GUI file manager or the terminal, you can have a brief look at the dataset and will see that it is a classic
+``BIDS`` dataset with the expected ``paths`` and ``files`` (which is obviously important given that ``BIDSonym`` is a ``BIDS App``).
+
+.. image:: https://raw.githubusercontent.com/PeerHerholz/BIDSonym/master/docs/source/_static/example_dataset.png
+    :alt: alternate text
+
+**2. Installing BIDSonym**
+
+Of course we need to install ``BIDSonym`` before we can run it. Following the `Installation instructions and guidelines <./installation.rst>`_,
+we all utilize the ``Docker image`` and thus a containerized version of ``BIDSonym``. Getting everything ready to run is fast and easy via:
+
+.. code-block:: bash
+
+    docker pull peerherholz/bidsonym
+
+This ``docker command`` will download the ``latest`` version of ``BIDSonym`` from `DockerHub <https://hub.docker.com/r/peerherholz/bidsonym>`_ and
+if everything worked as expected, you should see the following message:
+
+.. code-block:: bash
+
+    Status: Downloaded newer image for peerherholz/bidsonym:latest
+
+With that, we ready to test ``BIDSonym`` on the example dataset.
+
+**3. Running BIDSonym**
+
+Now that everything is in place, we can run ``BIDSonym`` on the example dataset.  
+After having a look at the `Usage information <./usage.rst>`_ again, we decide to
+utilize `pydeface <https://github.com/poldracklab/pydeface/>`_ as the ``defacing algorithm`` and use `bet <https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/BET>`_ for brain extraction
+prior to defacing (applying the default fractional intensity threshold of 0.5). Additionally, we want to delete the
+``meta-data information`` provided under ``InstitutionName``, ``InstitutionalDepartmentName`` and ``InstitutionAddress`` to remove 
+information that could be potentially helpful in re-identifying participants. As the example datasets only contains data from 
[…]
