### 2021-08-28 b327373
Update docs after building Travis build 159 of PeerHerholz/BIDSonym
The docs were built from the branch 'master' against the commit
057fc7c5fbefbcbcf2e641731f64afd2777d984d.

The Travis build that generated this commit is at
https://app.travis-ci.com/PeerHerholz/BIDSonym/jobs/534326931.

The doctr command that was run is

    /home/travis/virtualenv/python3.6.7/bin/doctr deploy --built-docs docs/build/html .
