### 2021-08-28 #3169 @PeerHerholz
Thanks for your help and effort @danielskatz!

I completed all the outlined steps and the [tagged release is now available on zenodo](https://zenodo.org/record/5301625#.YSoF29Mzblw), with the `DOI` being
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5301625.svg)](https://doi.org/10.5281/zenodo.5301625).

Thanks again, best, Peer
