### 2021-08-28 #3169 @danielskatz
I've suggested a bunch of changes to the paper wording in https://github.com/PeerHerholz/BIDSonym/pull/62 - Also in Figure 1,  I see a "meta-data" that should be "metadata" to be consistent with the rest of the paper.  Please merge the PR (or let me know what you disagree with), and let me know if you want to change Figure 1, then we can proceed to final acceptance and publication.
