### 2021-08-28 94f899c
update BIDSonym functionality graphic
This commit adapts the BIDSonym functionality graphic to change "meta-data" to "metadata" as suggested by @danielskatz [here](https://github.com/openjournals/joss-reviews/issues/3169#issuecomment-907606652).
