### 2021-08-28 #3169 @PeerHerholz
Hi @danielskatz,

thx so much for all of this! I accepted the PR and [adapted Figure 1](https://github.com/PeerHerholz/BIDSonym/commit/94f899c5fa3cf62460e9386d34453565e757d03f).

Best, Peer
