### 2021-08-28 #3169 @danielskatz
@also, please remove the abstract field from any references that do not have DOIs, as these get copied into unstructured metadata that is sent to crossref for for such references, which isn't needed.
