### 2021-08-28 #3169 @danielskatz
Also, in the caption of Figure 1, there's one "meta-data" left that I missed - this should be changed to "metadata"
