### 2021-08-28 #3169 @danielskatz
And for the McKinney reference in the bib file, please change the "Python" in the title to "{P}ython" to preserve the correct case for the P.
