+++ paper/paper.md
@@ -16,7 +16,7 @@ authors:
     orcid: 0000-0002-9794-749X
     affiliation: 1
 affiliations:
- - name: NeuroDataScience-Origami lab, McConnell Brain Imaging Centre, The Neuro (Montreal Neurological Institute-Hospital), Faculty of Medicine, McGill
+ - name: NeuroDataScience:ORIGAMI lab, McConnell Brain Imaging Centre, The Neuro (Montreal Neurological Institute-Hospital), Faculty of Medicine, McGill
           University, Montreal, Quebec, Canada
    index: 1
  - name: University of Oregon, Eugene, United States
@@ -30,8 +30,7 @@ bibliography: BIDSonym.bib
 Due to the evolution of research incentives, technical advancements, and the development of new standards [@eickhoff_sharing_2016; @gorgolewski_brain_2016; @nichols_best_2017; @poldrack_toward_2013; @poldrack_making_2014; @poldrack_openfmri_2017], increasingly greater amounts of neuroimaging data are being shared either publicly or made available through data user agreements. These datasets originate from small samples of participants collected by individual research groups, as well as from “Big Data” samples including thousands of participants collected by large research consortia (UK Biobank [@sudlow_uk_2015], HCP [@van_essen_wu-minn_2013], ABIDE [@di_martino_autism_2014], ADNI [@mueller_alzheimers_2005], etc.) While data sharing is important and beneficial [@eickhoff_sharing_2016; @nichols_best_2017; @poldrack_making_2014; @poline_data_2012], the privacy of participant data must be protected [@bannier_open_2020; @brakewood_ethics_2013]. To that end, Ethic Review Boards and data sharing platforms typically require that uploaded datasets are provided in anonymized or pseudo-anonymized form, limiting participant reidentification.  However, the (pseudo-)anonymization process is deceptively complex; attempts at ensuring data privacy must take into consideration all dataset components, including imaging modalities, as well as national legal and ethical frameworks. Several algorithms have been developed to (pseudo-)anonymize imaging datasets but they offer limited solutions. Some are attached to specific software and some are limited to specific computing environments; most miss an in-depth assessment and treatment of the metadata attached to the dataset or lack the capacity to automatize (pseudo-)anonymization across large datasets. BIDSonym was created to address these points in one simple, flexible, and general tool that offers users an array of automated (pseudo-)anonymization options to augment participant privacy in neuroimaging datasets. There are two components of neuroimaging datasets that arguably pose the largest risk to maintaining participant privacy: the structural images and accompanying metadata (e.g., metadata text files or information embedded in image file headers). Structural images contain visible identifiable participant information via facial features like the eyes, nose, and mouth, and privacy is usually addressed through a process called “defacing”, within which all or a subset of these features are removed from the final structural data files. The metadata text files may additionally contain identifiable participant data through the recording of acquisition time and location, and personal details such as date of birth, height, and weight. Here, privacy is maintained by removing or blurring this information from the final dataset. BIDSonym addresses both vulnerabilities in neuroimaging datasets, obviating the need for multiple steps within a data sharing pipeline to ensure participant privacy.
 
 ## Summary
-In concordance with the BIDS-App template [@gorgolewski_bids_2017], BIDSonym operates as a command line tool written in Python [@rossum_python_1995] and
[…]
