+++ paper/BIDSonym.bib
@@ -265,7 +265,6 @@
 	address = {NLD},
 	type = {Technical {Report}},
 	title = {Python reference manual},
-	abstract = {Python is a simple, yet powerful, interpreted programming language that bridges the gap between C and shell programming, and is thus ideally suited for ``throw-away programming'''' and rapid prototyping. Its syntax is put together from constructs borrowed from a variety of other languages; most prominent are influences from ABC, C, Modula-3 and Icon. The Python interpreter is easily extended with new functions and data types implemented in C. Python is also suitable as an extension language for highly customizable C applications such as editors or window managers. Python is available for various operating systems, amongst which several flavors of UNIX (including Linux), the Apple Macintosh O.S., MS-DOS, MS-Windows 3.1, Windows NT, and OS/2. This reference manual describes the syntax and ``core semantics'''' of the language. It is terse, but attempts to be exact and complete. The semantics of non-essential built-in object types and of the built-in functions and modules are described in the Python Library Reference. For an informal introduction to the language, see the Python Tutorial.},
 	institution = {CWI (Centre for Mathematics and Computer Science)},
 	author = {Rossum, Guido},
 	year = {1995}
@@ -301,7 +300,6 @@
 	title = {mridefacer},
 	copyright = {View license         ,                 View license},
 	url = {https://github.com/mih/mridefacer},
-	abstract = {Helper to aid de-identification of MRI images (3D or 4D)},
 	urldate = {2020-12-24},
 	author = {Hanke, Michael and Halchenko, Yaroslav},
 	year = {2018},
@@ -312,7 +310,6 @@
 	address = {USA},
 	series = {{HealthSec}'11},
 	title = {Quickshear defacing for neuroimages},
-	abstract = {Data sharing offers many benefits to the neuroscience research community. It encourages collaboration and interorganizational research efforts, enables reproducibility and peer review, and allows meta-analysis and data reuse. However, protecting subject privacy and implementing HIPAA compliance measures can be a burdensome task. For high resolution structural neuroimages, subject privacy is threatened by the neuroimage itself, which can contain enough facial features to re-identify an individual. To sufficiently de-identify an individual, the neuroimage pixel data must also be removed. Quickshear Defacing accomplishes this task by effectively shearing facial features while preserving desirable brain tissue.},
 	urldate = {2020-12-24},
 	booktitle = {Proceedings of the 2nd {USENIX} conference on {Health} security and privacy},
 	publisher = {USENIX Association},
@@ -388,10 +385,11 @@
 }
 
 @inproceedings{mckinney_data_2010,
-	title = {Data structures for statistical computing in python},
+	title = {Data structures for statistical computing in {P}ython},
 	volume = {445},
 	booktitle = {Proceedings of the 9th {Python} in {Science} {Conference}},
 	author = {McKinney, Wes},
 	year = {2010},
-	pages = {51--56}
+	pages = {51--56},
+	doi = {10.25080/Majora-92bf1922-00a}
 }
