### 2021-08-28 #3169 @PeerHerholz
Hi @danielskatz,

sorry for all these missed little things.

I added the missing `DOI`, removed the `abstract` field from references that don't have a `DOI`, adapted the caption and changed the title of the McKinney reference. I hope everything is ok now.

Thanks again, best, Peer
